# Improve performance of `Arr::dot([])`