@browner12 2023-12-14
Wouldn't array unpacking via the splat operator be more appropriate here to maintain identical behavior?  I remember hearing awhile ago that unpacking was more performant than `array_merge()` but don't know if that still holds true.