@mferrara 2023-12-14
> I don't understand some of these changes. Why is there a semicolon after the `foreach` loop?

Oof. Guess that's that.