@browner12 2023-12-14
you could try again, but limit it to only making the desired change on line 116.  don't add comments, don't change spacing, don't add unnecessary semicolons. that's gonna get you a much higher change of this being considered.

I'd also run and include some benchmarks against array unpacking, because the behavior change of the union operator might be a tough pill to swallow.  If you can include your script for *how* you ran the benchmark, that can also be helpful.